% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor2.R
\name{cor2}
\alias{cor2}
\title{Correlation Analysis}
\usage{
cor2(data, x1, x2, method = "spearman")
}
\arguments{
\item{data}{dataframe}

\item{x1}{x1}

\item{x2}{x2}

\item{method}{1,2,3}
}
\description{
Correlation Analysis
}
