% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_sig.R
\name{cor_sig}
\alias{cor_sig}
\title{Correlation Analysis with Signicant Values}
\usage{
cor_sig(data, method = "pearson")
}
\arguments{
\item{data}{a dataframe or matrix}

\item{method}{a character string indicating which correlation coefficient (or covariance) is to be computed. One of "pearson" (default), "kendall", or "spearman": can be abbreviated.}
}
\value{
correlation analysis with significant star.
}
\description{
Correlation Analysis with Signicant Values
}
\examples{
cor_sig(mtcars)
}
