% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_star.R
\name{cor_star}
\alias{cor_star}
\title{Correlation Analysis}
\usage{
cor_star(data, method = "pearson")
}
\arguments{
\item{data}{a dataframe or matrix}

\item{method}{a character string indicating which correlation coefficient (or covariance) is to be computed. One of "pearson" (default, 1), "kendall" (2), or "spearman" (3): can be abbreviated.}
}
\value{
correlation analysis with significant star.
}
\description{
Correlation Analysis
}
\examples{
cor_star(mtcars)
}
