\name{cucconi-package}
\alias{cucconi-package}
\alias{cucconi}
\docType{package}
\title{
  Multisample Cucconi Test
}
\description{
  An implementation of the Multisample Cucconi Test.
}
\details{
  This section should provide a more detailed overview of how to use the
  package, including the most important functions.
}
\author{
Filippo Maganza

Maintainer: Filippo Magnza <your@email.com>
}
\references{
  This optional section can contain literature or other references for
  background information.
}
\keyword{ package }
\seealso{
  Optional links to other man pages
}
\examples{
  \dontrun{
     ## Optional simple examples of the most important functions
     ## These can be in \dontrun{} and \donttest{} blocks.   
  }
}
