% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/digital.R
\name{digital}
\alias{digital}
\title{Set Digital Number}
\usage{
digital(x, round)
}
\arguments{
\item{x}{vector, dataframe or matrix}

\item{round}{digital number}
}
\value{
character with the same digital number
}
\description{
Set Digital Number
}
\examples{
digital(1.2,4)
}
