% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.factor.R
\name{list.factor}
\alias{list.factor}
\title{Return All Factor Variables}
\usage{
list.factor(x, levels = FALSE)
}
\arguments{
\item{x}{a dataframe or matrix}

\item{levels}{logical. TRUE to display levels for factor variable.}
}
\value{
factor variable names and levels
}
\description{
Return all factor variables in a dataframe or matrix
}
\examples{
jh=data.frame(x=c(1,2,3,1),
    k=c(4,5,6,7),
    h=c('a','a','b','b'))
list.factor(jh)
}
