% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.NA.R
\name{list.NA}
\alias{list.NA}
\title{Return Na Count and Percentage}
\usage{
list.NA(x)
}
\arguments{
\item{x}{a numeric vector, a dataframe or matix}
}
\value{
A dataframe contains NA variable names, NA count and percentage
}
\description{
Return Na count and percentage for each variable in a dataframe or matrix.
}
\examples{
 jh=data.frame(x=c(1,2,3,1),
     k=c(4,5,6,7),
     h=c('a','a',NA,'D'),
     f=c(1,2,NA,NA))
list.NA(jh)
}
