% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.numeric.R
\name{list.numeric}
\alias{list.numeric}
\title{Return All Numeric Variables in A Dataframe}
\usage{
list.numeric(df)
}
\arguments{
\item{df}{a dataframe}
}
\value{
numeric variable names
}
\description{
Return All Numeric Variables in A Dataframe
}
\examples{
jh=data.frame(x=c(1,2,3,1),
    k=c(4,5,6,7),
    h=c('a','a','b','b'))
list.numeric(jh)
}
