% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.str.R
\name{list.str}
\alias{list.str}
\title{Structure for Data}
\usage{
list.str(x, n = 3)
}
\arguments{
\item{x}{a dataframe or matrix}

\item{n}{the maximum level number to display}
}
\value{
a dataframe contains variable names and class
}
\description{
Structure for Data
}
\examples{
jh=data.frame(x=c(1,2,3,1),
              k=c(4,5,6,7),
              h=c('a','a','b','b'))
list.str(x = jh)
}
