% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.summary.R
\name{list.summary}
\alias{list.summary}
\title{Summary for Data}
\usage{
list.summary(x, round = 2)
}
\arguments{
\item{x}{numeric}

\item{round}{digital number}
}
\value{
a dataframe with min, max, quantile 25 and 75, mean, median, sd and NA
}
\description{
Summary for Data
}
\examples{
list.summary(mtcars)
}
