% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mv_linear.R
\name{mv_linear}
\alias{mv_linear}
\title{Multivariable Linear Regression}
\usage{
mv_linear(data, y, x, direction = "no", summary = TRUE, ...)
}
\arguments{
\item{data}{data}

\item{y}{y varaible}

\item{x}{variable names for univariable linear regression. If missing, it will be column names of data except y and adjust}

\item{direction}{direction for stepwise regression. Four options: no, backward, forward and both. Defaulted is no}

\item{summary}{logical. Whether to return summary results. TRUE as defaulted}

\item{...}{arguments passed to step() function}
}
\value{
multivariable linear regression results
}
\description{
Multivariable Linear Regression
}
\examples{
mv_linear(data = rock,y = 'perm',
    direction = 'both')
}
