% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mv_logit.R
\name{mv_logit}
\alias{mv_logit}
\title{Multivariable Logistic Regression}
\usage{
mv_logit(data, y, x, direction = "no", summary = TRUE, ...)
}
\arguments{
\item{data}{data}

\item{y}{y varaible}

\item{x}{variable names for univariable logistic regression. If missing, it will be column names of data except y and adjust}

\item{direction}{direction for stepwise regression. Four options: no, backward, forward and both. Defaulted is no}

\item{summary}{logical. Whether to return summary results. TRUE as defaulted}

\item{...}{arguments passed to step() function}
}
\value{
multivariable logistic regression results
}
\description{
Multivariable Logistic Regression
}
\examples{
mv_logit(data = mtcars,y = 'am',
    variable = c('cyl','disp'))
}
