% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round_set.R
\name{round<-}
\alias{round<-}
\title{Change the Digital for Double}
\usage{
round(x) <- value
}
\arguments{
\item{x}{a double number}

\item{value}{digital number}
}
\value{
double number
}
\description{
Change the Digital for Double
}
\examples{
x = 3.123
#usual method
x = round(x, 3)
#now
round(x) = 3
}
