% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surdiff_p.value.R
\name{survdiff_p.value}
\alias{survdiff_p.value}
\title{Extract P Value after survdiff() function}
\usage{
survdiff_p.value(survdiff)
}
\arguments{
\item{survdiff}{the results of survdiff() function}
}
\value{
p value
}
\description{
Extract P Value after survdiff() function
}
\examples{
library(survival)
diff_result=survdiff(Surv(qsec,vs)~cyl,data=mtcars)
survdiff_p.value(diff_result)
}
