% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survSum.R
\name{survSum}
\alias{survSum}
\alias{surv_table}
\alias{surv_median_time}
\alias{surv_year_rate}
\title{Calculate Survival Rate and Time}
\usage{
surv_table(fit)

surv_median_time(fit)

surv_year_rate(fit, year)
}
\arguments{
\item{fit}{fit by survfit() function}

\item{year}{year}
}
\value{
a dataframe
}
\description{
Calculate Survival Rate and Time
}
\examples{
library(survival)
fit=survfit(Surv(futime, fustat)~rx, data=ovarian)

#survival table
surv_table(fit)

#median survival rate
surv_median_time(fit)

#one year survaival rate
surv_year_rate(fit,365)

#two years survaival rate
surv_year_rate(fit,365*2)

}
