% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to.factor-set.R
\name{to.factor<-}
\alias{to.factor<-}
\title{Set Factor Class}
\usage{
to.factor(x) <- value
}
\arguments{
\item{x}{the data that you want to set}

\item{value}{levels, the first value is the reference. If the length of value is 1, no levels will be given to x}
}
\value{
factor x
}
\description{
Set Factor Class
}
\examples{
to.factor(mtcars$gear) <- c(4,3,5)
}
