% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to.factor.R
\name{to.factor}
\alias{to.factor}
\title{Set Factor Class}
\usage{
to.factor(x, levels)
}
\arguments{
\item{x}{the data that you want to set}

\item{levels}{levels, the first levels is the reference. If the length of levels is 1, no levels will be given to x}
}
\value{
factor x
}
\description{
Set Factor Class
}
\examples{
to.factor(mtcars$gear,c(4,3,5))
to.factor(mtcars$gear)
}
