% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to.numeric.R
\name{to.numeric}
\alias{to.numeric}
\title{Change to Numeric Form}
\usage{
to.numeric(x)
}
\arguments{
\item{x}{vector}
}
\value{
numeric data
}
\description{
Change to Numeric Form
}
\examples{
x=c(1,2,3)
to.factor(x) <- 1
to.numeric(x)
}
