% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to.refer-set.R
\name{to.refer<-}
\alias{to.refer<-}
\title{Set Refer for Factor}
\usage{
to.refer(x) <- value
}
\arguments{
\item{x}{the data that you want to set}

\item{value}{refering level}
}
\value{
refered factor value
}
\description{
Convert data to be factor and set reference.
}
\examples{
to.refer(mtcars$vs) = 1
}
