% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to.refer.R
\name{to.refer}
\alias{to.refer}
\title{Set Refer for Factor}
\usage{
to.refer(x, refer)
}
\arguments{
\item{x}{the data that you want to set}

\item{refer}{refering level}
}
\value{
refered factor refer
}
\description{
Convert data to be factor and set reference.
}
\examples{
to.refer(mtcars$vs,1)
}
