% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uv_logrank.R
\name{uv_logrank}
\alias{uv_logrank}
\title{Looping for logrank Regression}
\usage{
uv_logrank(data, time, event, variable, round = 3, order_by.p = TRUE)
}
\arguments{
\item{data}{data}

\item{time}{time variable}

\item{event}{event variable}

\item{variable}{variable names for logrank regression. If missing, it will be column names of data except y}

\item{round}{digital round, 3 is defaulted}

\item{order_by.p}{logical. TRUE, defaulted, means increasing order in p value}
}
\value{
logrank regression results
}
\description{
Looping for logrank Regression
}
\examples{
uv_logrank(data = mtcars,
    time = 'qsec',event = 'vs')
}
