% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.plot.R
\name{list.plot}
\alias{list.plot}
\title{Scatter Plot for Single Value}
\usage{
list.plot(x, label = "x")
}
\arguments{
\item{x}{vector, dataframe or matrix}

\item{label}{labels for points. If missing, defaulted, no labels will be added. If label equals x, id will be added. If label equals y, y value will be added.}
}
\value{
sactter
}
\description{
Scatter Plot for Single Value
}
\examples{
\donttest{
    list.plot(mtcars)
}

}
